[#57] Remove `ProjectLinking` and inject current package and imports into contexts
Currently, the compiler build the pure AST, then inject full types in it.
When coming to contexts, the plugin cannot know its own package, nor its imports.

We change the strategy by removing the `ProjectLinking`.
Doing so, we need to inject full type names as soon as possible
and extract current type names before building pure root.

This commit does the following:

* Remove `ProjectLinking`
* Create `DefinitiFileASTParser` parse a file with package and import information
* Move some methods from `CommonParser` to `DefinitiFileASTParser`*
* Extract current file names to define the current file type imports as it was done with `ProjectLinking`
* Update the `ContextPlugin` to inject package name and imports when building context AST
* Update rest of code to be compatible with the new structure
