--- src/main/scala/definiti/core/linking/package.scala
@@ -1,7 +1,0 @@
-package definiti.core
-
-package object linking {
-  private[core] type TypeMapping = Map[String, String]
-
-  private[core] def emptyTypeMapping = Map.empty[String, String]
-}
