+++ src/main/scala/definiti/core/parser/project/CommonParser.scala
@@ -11,48 +11,6 @@ import scala.collection.mutable.ListBuffer
 private[core] trait CommonParser {
   def file: String
 
-  def processParameterListDefinition(context: ParameterListDefinitionContext): Seq[ParameterDefinition] = {
-    scalaSeq(context.parameterDefinition()).map(processParameter)
-  }
-
-  def processParameter(context: ParameterDefinitionContext): ParameterDefinition = {
-    ParameterDefinition(
-      name = context.parameterName.getText,
-      typeReference = processTypeReference(context.typeReference()),
-      location = Location(file, getRangeFromContext(context))
-    )
-  }
-
-  def processTypeReference(context: TypeReferenceContext): TypeReference = {
-    TypeReference(context.name.getText, processGenericTypeList(context.genericTypeList()))
-  }
-
-  def processGenericTypeList(context: GenericTypeListContext): Seq[TypeReference] = {
-    if (context != null) {
-      scalaSeq(context.genericType()).map(processGenericType)
-    } else {
-      Seq()
-    }
-  }
-
-  def processGenericType(context: GenericTypeContext): TypeReference = {
-    TypeReference(
-      context.IDENTIFIER().getText,
-      processGenericTypeList(context.genericTypeList())
-    )
-  }
-
-  def extractStringContent(string: String): String = {
-    var temporaryResult = string
-    if (temporaryResult.startsWith("\"")) {
-      temporaryResult = temporaryResult.substring(1)
-    }
-    if (temporaryResult.endsWith("\"")) {
-      temporaryResult = temporaryResult.substring(0, temporaryResult.length - 1)
-    }
-    temporaryResult
-  }
-
   def extractDocComment(string: String): String = {
     var temporaryResult = string
     if (temporaryResult.startsWith("/**")) {
