+++ src/test/scala/definiti/core/mock/plugins/StringExtendedContext.scala
@@ -7,7 +7,7 @@ import definiti.core.{ContextPlugin, Valid, Validated}
 class StringExtendedContext extends ContextPlugin[String] {
   override def contextName = "stringContext"
 
-  override def parse(content: String, location: Location): String = content
+  override def parse(content: String, packageName: String, imports: Map[String, String], location: Location): String = content
 
   override def validate(context: String, library: Library): Validated[NoResult] = Valid(NoResult)
